requirements.txt
numpy
Pillow
protobuf
six
tensorboardX
torch
torchvision
tqdm
pandas
h5py
matplotlib
xarray
netCDF4
scipy
